Prototype for a path finder in C++
Everything useful is in Graph.cpp and is exported through Graph.h.
Dijkstra.cpp is just a main function that uses it, it'll probably vanish
soon. The code isn't particularly optimised so I can almost certainly
make it run faster if required although it'll only get less readable. My
intention after I've cleaned the code up a bit and done some testing is
to wrap the whole thing in python's foreign function interface. The end
result should be an object that looks like a module to the rest of the
code but is implemented in C++.